Restructure the repo for qol.
